components/rules_engine/requirements.txt
fastapi-restful==0.4.2
fastapi==0.85.1
fireo==1.6.2
google-cloud-bigquery==3.10.0
google-cloud-storage==2.10.0
python-multipart==0.0.5
regex==2023.10.3
uvicorn==0.19.0
zen-engine==0.15.0
